Add CVC -> Google Sheets scraper
Add a script that maintains the spreadsheet at
https://docs.google.com/spreadsheets/d/1NR36K5nBy4rI69qU6dc5MGt2QwwTXZugDnjlWnei3X8/edit#gid=19019745.
It creates one sheet inside per state, and one row per CVC.
Each column represents a date, and indicates the number of
vaccinations that have happened there on that date.

Scraping is done concurrently to speed things up. Can be
made faster if running on a server, since currently network
bandwidth is the bottleneck.

Needs better error handling and retry logic, however.